split GroupNotificationService into multiple smaller services to avoid dependency hell
