+++ src/test/java/de/tum/in/www1/artemis/service/notifications/GroupNotificationServiceTest.java
@@ -46,6 +46,9 @@ class GroupNotificationServiceTest extends AbstractSpringIntegrationBambooBitbuc
     @Autowired
     private ExamRepository examRepository;
 
+    @Autowired
+    private GroupNotificationScheduleService groupNotificationScheduleService;
+
     private Exercise exercise;
 
     private Exercise updatedExercise;
@@ -221,7 +224,7 @@ class GroupNotificationServiceTest extends AbstractSpringIntegrationBambooBitbuc
      */
     @Test
     void testCheckNotificationForExerciseRelease_undefinedReleaseDate() {
-        groupNotificationService.checkNotificationsForNewExercise(exercise);
+        groupNotificationScheduleService.checkNotificationsForNewExercise(exercise);
         verify(groupNotificationService, times(1)).notifyAllGroupsAboutReleasedExercise(any());
     }
 
@@ -231,7 +234,7 @@ class GroupNotificationServiceTest extends AbstractSpringIntegrationBambooBitbuc
     @Test
     void testCheckNotificationForExerciseRelease_currentOrPastReleaseDate() {
         exercise.setReleaseDate(CURRENT_TIME);
-        groupNotificationService.checkNotificationsForNewExercise(exercise);
+        groupNotificationScheduleService.checkNotificationsForNewExercise(exercise);
         verify(groupNotificationService, times(1)).notifyAllGroupsAboutReleasedExercise(any());
     }
 
@@ -241,7 +244,7 @@ class GroupNotificationServiceTest extends AbstractSpringIntegrationBambooBitbuc
     @Test
     void testCheckNotificationForExerciseRelease_futureReleaseDate() {
         exercise.setReleaseDate(FUTURE_TIME);
-        groupNotificationService.checkNotificationsForNewExercise(exercise);
+        groupNotificationScheduleService.checkNotificationsForNewExercise(exercise);
         verify(instanceMessageSendService, times(1)).sendExerciseReleaseNotificationSchedule(any());
     }
 
@@ -258,7 +261,7 @@ class GroupNotificationServiceTest extends AbstractSpringIntegrationBambooBitbuc
         updatedExercise.setReleaseDate(dateOfUpdatedExercise);
         updatedExercise.setAssessmentDueDate(dateOfUpdatedExercise);
 
-        groupNotificationService.checkAndCreateAppropriateNotificationsWhenUpdatingExercise(exercise, updatedExercise, NOTIFICATION_TEXT);
+        groupNotificationScheduleService.checkAndCreateAppropriateNotificationsWhenUpdatingExercise(exercise, updatedExercise, NOTIFICATION_TEXT);
 
         verify(groupNotificationService, times(1)).notifyAboutExerciseUpdate(any(), any());
